renamed smc to stateMachine
